+++ help/manifest.json
@@ -1,7 +1,7 @@
 {
   "project": "MissionChief Map Command Toolkit",
-  "guideVersion": "8.2.6",
-  "toolkitVersion": "8.2.6",
+  "guideVersion": "8.2.7",
+  "toolkitVersion": "8.2.7",
   "updated": "2026-07-29",
   "path": "help/index.html",
   "sections": 21,
@@ -15,6 +15,6 @@
   "financialIntelligenceRules": "financial-intelligence/v1/classification-rules.json",
   "financialAuditPolicy": "financial-intelligence/v2/audit-policy.json",
   "financialArchitectureManifest": "financial-intelligence/v2/manifest.json",
-  "runtimeGuidePatch": "Toolkit v8.2.6 restores the native mission \u2192 FMS 5 vehicle \u2192 Discharge patient \u2192 Patient isn\u2019t transported workflow.",
+  "runtimeGuidePatch": "Toolkit v8.2.7 recognises and clicks the real Cancel Transport link/button and waits for Patient isn’t transported or a genuine native transition.",
   "godfatherThemePackage": "themes/godfather/manifest.json"
 }
